requirements.txt
Flask==2.3.2
python-dotenv==1.0.0
Werkzeug==2.3.7
bcrypt==4.0.1
PyJWT==2.8.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
gunicorn==22.0.0